Rename ThrowExtensions.h to ExceptionFactory.h in C++ code
The functions in the C++ version return exception objects rather than throwing them,
making "ThrowExtensions" a misleading name. "ExceptionFactory" better reflects that
these are factory functions that create exception objects for later use.

Updated all references:
- Platform.Exceptions.h include statement
- Visual Studio project files (.vcxproj and .vcxproj.filters)

Fixes #76
