+++ cpp/Platform.Exceptions/ExceptionFactory.h
